Eliminate "globals" au8Buffer and u8BufferSize.
The goal here is to make refactoring easier by cutting out
global (shared) variables, and making functions self-contained.
Functions with "side-effects" make code more highly-connected
than it needs to be.

The ideal is to pass in everything needed by the function as
parameters.

This also helps to cut down on memory use. The working buffer is
now on the stack, so that memory is freed up for other code to
use, when we're not inside a MODBUS function call.

For new parameters, I've not stuck to the type-prefix notation.
This keeps variable names distinct, and helps make sure that
there are no stray references to global variables.