Split class Modbus into class Master & class Slave
The whole library has been placed into a namespace ("modbus").
By default the namespace is "used", so existing code will work.
But the increasing number of names may eventually clash with
someone else's code. Set macro USING_MODBUS_NAMESPACE to 0 to
force the namespace to be explicit.

There is a shared base class "Base" which encapsulates comms-
related functionality.

I've retained the class "Modbus", but now it's just a shim, which
creates a modbus::Master or modbus::Slave object, as necessary,
and passes on method-calls to it.

This change helps to disentangle the master and slave code from
each other. It also saves memory, by removing unused member
variables.